suspenseをSSG状態でも使えるように変更
